+++ src/include/day.h
@@ -1,7 +1,8 @@
 #ifndef H_DAY
 #define H_DAY
+#include <stdint.h>
 #include <stdlib.h>
 
-int part_1(char* input, size_t strlen);
-int part_2(char* input, size_t strlen);
+uint64_t part_1(char* input, size_t strlen);
+uint64_t part_2(char* input, size_t strlen);
 #endif
